Refactor no-missing-import, no-unclosed-tag, no-unknown-tag-name and no-unknown-slot into rule modules
